Minor refactor.
Moved common functions to util.
